requirements.txt
Django==4.2.9
django-taggit
django-htmx
Markdown